requirements.txt
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
sqlalchemy>=2.0.14,<3
alembic>=1.13,<2
psycopg[binary]>=3.1,<4
httpx>=0.26,<1
pydantic-settings>=2.1,<3
apscheduler>=3.10,<4
